chore(core): extract ProgressTopics into utils/progress-topics
The topic constant is the contract between progress producers
(DelayedSpinner callers) and the daemon subscriber registry, so it
belongs next to the spinner rather than inside the daemon's
client-socket bookkeeping. Also sidesteps a latent import cycle
between delayed-spinner.ts and the daemon server module.

https://claude.ai/code
